Refactor API controllers and services, update validation rules, and enhance response handling for client packages, limits, and items

+++ routes/api.php
@@ -29,42 +29,27 @@ Route::middleware([
 Route::middleware('auth:sanctum')->group(function () {
 
     //Chats
-    //step 3
     Route::post('/chats', [ChatController::class, 'createChat'])->name('chat')->middleware('role:admin');
-
-    //step none
     Route::delete('/chats/{chatId}', [ChatController::class, 'deleteChat'])->name('deletechat')->middleware('role:admin');
-
-    //step 3
     Route::get('/chats', [ChatController::class, 'getUserChats'])->name('getchats');
-
-    //step none
     Route::get('/chats/{chatId}', [ChatController::class, 'getUserChat'])->name('getchat');
 
 
     //Messages
-
-    //step 7  (store)
     Route::post('/chats/{chatId}/messages', [ChatMessageController::class, 'sendMessage'])->name('sendmessage');
-
-    //step 8  (index)
     Route::get('/chats/{chatId}/messages', [ChatMessageController::class, 'getMessages'])->name('getmessages');
 
 
     //Team
-    //step 4
     Route::post('/chats/{chatId}/assign-team', [ChatController::class, 'assignTeam'])->name('assignteam')->middleware('role:admin');
 
 
-    //step 6
     //Package - Chat
     Route::post('/chats/{chatId}/attach-package', [ChatController::class, 'attachPackage'])->name('attachpackage')->middleware('role:admin');
 
 
 
-
     // Package Routes
-
     Route::prefix('packages')->group(function () {
         Route::get('/', [PackageController::class, 'index'])->name('packages.index')->middleware('role:admin|team');
         Route::get('/{packageId}', [PackageController::class, 'show'])->name('packages.show')->middleware('role:admin|team');
@@ -74,6 +59,7 @@ Route::middleware('auth:sanctum')->group(function () {
     });
 
 
+    //Package Items
     Route::prefix('package-items')->group(function () {
         Route::get('/{packageId}', [PackageItemController::class, 'index'])->name('package-items.index')->middleware('role:admin|team');
         Route::post('/', [PackageItemController::class, 'store'])->name('package-items.store')->middleware('role:admin');
@@ -83,47 +69,59 @@ Route::middleware('auth:sanctum')->group(function () {
     });
 
 
+    //Item Types
     Route::prefix('item-types')->group(function () {
-        Route::get('/', [ItemTypeController::class, 'index']);
-        Route::post('/', [ItemTypeController::class, 'store']);
-        Route::get('/{id}', [ItemTypeController::class, 'show']);
-        Route::put('/{id}', [ItemTypeController::class, 'update']);
-        Route::delete('/{id}', [ItemTypeController::class, 'destroy']);
+        Route::get('/', [ItemTypeController::class, 'index'])->name('item-types.index')->middleware('role:admin|team');
+        Route::post('/create', [ItemTypeController::class, 'store'])->name('item-types.store')->middleware('role:admin');
+        Route::get('/{id}', [ItemTypeController::class, 'show'])->name('item-types.show')->middleware('role:admin|team');
+        Route::put('/{id}', [ItemTypeController::class, 'update'])->name('item-types.update')->middleware('role:admin');
+        Route::delete('/{id}', [ItemTypeController::class, 'destroy'])->name('item-types.destroy')->middleware('role:admin');
     });
 
 
+    //Package Allowed Items
     Route::prefix('package-allowed-items')->group(function () {
-        Route::get('/', [PackageAllowedItemController::class, 'index']);
-        Route::get('/{id}', [PackageAllowedItemController::class, 'show']);
-        Route::post('/', [PackageAllowedItemController::class, 'store']);
-        Route::put('/{id}', [PackageAllowedItemController::class, 'update']);
-        Route::delete('/{id}', [PackageAllowedItemController::class, 'destroy']);
-        Route::get('/by-package/{packageId}', [PackageAllowedItemController::class, 'byPackage']);
+        Route::get('/', [PackageAllowedItemController::class, 'index'])->name('package-allowed-items.index')->middleware('role:admin|team');
+        Route::get('/{id}', [PackageAllowedItemController::class, 'show'])->name('package-allowed-items.show')->middleware('role:admin|team');
+        Route::post('/store', [PackageAllowedItemController::class, 'store'])->name('package-allowed-items.store')->middleware('role:admin');
+        Route::put('/{id}', [PackageAllowedItemController::class, 'update'])->name('package-allowed-items.update')->middleware('role:admin');
+        Route::delete('/{id}', [PackageAllowedItemController::class, 'destroy'])->name('package-allowed-items.destroy')->middleware('role:admin');
     });
 
 
     ////////////////////////////////////////////////////////////////////////////////////////////
 
-
+    //Client Package
     Route::prefix('client-package')->group(function () {
-        Route::post('log-usage', [ClientPackageController::class, 'logUsage']);
-        Route::post('change-status', [ClientPackageController::class, 'changeStatus']);
-        Route::get('/', [ClientPackageController::class, 'index']);
-        Route::get('{clientPackageId}', [ClientPackageController::class, 'show']);
-        Route::post('/', [ClientPackageController::class, 'store']);
-        Route::patch('{clientPackageId}', [ClientPackageController::class, 'update']);
-        Route::delete('{clientPackageId}', [ClientPackageController::class, 'destroy']);
+        Route::post('/log-usage', [ClientPackageController::class, 'logUsage'])->name('client-package.log-usage')->middleware('role:admin');
+        Route::post('/change-status', [ClientPackageController::class, 'changeStatus'])->name('client-package.change-status')->middleware('role:admin');
+        Route::get('/', [ClientPackageController::class, 'index'])->name('client-package.index')->middleware('role:admin');
+        Route::get('/{clientPackageId}', [ClientPackageController::class, 'show'])->name('client-package.show')->middleware('role:admin');
+        Route::post('/', [ClientPackageController::class, 'store'])->name('client-package.store')->middleware('role:admin');
+        Route::patch('/{clientPackageId}', [ClientPackageController::class, 'update'])->name('client-package.update')->middleware('role:admin');
+        Route::delete('/{clientPackageId}', [ClientPackageController::class, 'destroy'])->name('client-package.destroy')->middleware('role:admin');
     });
 
-
+    //Client Package Items
     Route::prefix('client-package-items')->group(function () {
-        Route::put('{id}/edit', [ClientPackageItemController::class, 'editItem']);
-        Route::put('{id}/decline', [ClientPackageItemController::class, 'declineItem']);
-        Route::put('{id}/accept', [ClientPackageItemController::class, 'acceptItem']);
+        Route::get('/{clientPackageId}', [ClientPackageItemController::class, 'index']);
+        Route::get('/{clientPackageId}/{id}', [ClientPackageItemController::class, 'show']);
+        Route::post('/store/{clientPackageId}', [ClientPackageItemController::class, 'store']);
+        Route::patch('/{clientPackageId}/{id}', [ClientPackageItemController::class, 'update']);
+        Route::delete('/{clientPackageId}/{id}', [ClientPackageItemController::class, 'destroy']);
+        Route::put('{id}/accept', [ClientPackageItemController::class, 'accept']);
+        Route::put('{id}/edit', [ClientPackageItemController::class, 'edit']);
+        Route::put('{id}/decline', [ClientPackageItemController::class, 'decline']);
     });
 
 
+    //Client Limits
     Route::prefix('client-limits')->group(function () {
+        Route::get('/', [ClientLimitController::class, 'index']);
+        Route::post('/store', [ClientLimitController::class, 'store']);
+        Route::put('edit/{id}', [ClientLimitController::class, 'update']);
+        Route::delete('/{id}', [ClientLimitController::class, 'destroy']);
+
         Route::get('/{clientPackageId}', [ClientLimitController::class, 'remainingLimits']);
         Route::post('/{clientPackageId}/decrement-edit', [ClientLimitController::class, 'decrementEdit']);
         Route::post('/{clientPackageId}/decrement-decline', [ClientLimitController::class, 'decrementDecline']);
@@ -132,30 +130,32 @@ Route::middleware('auth:sanctum')->group(function () {
 
     /////////////////////////////////////////////////////////////////////////////////////
 
-
+    //Bonus Items
     Route::prefix('package-bonus-items')->group(function () {
         Route::get('bonus-items/{clientPackageId}', [BonusItemController::class, 'index']);
         Route::post('bonus-items', [BonusItemController::class, 'store']);
         Route::get('bonus-item/{id}', [BonusItemController::class, 'show']);
-        Route::put('bonus-item/{id}', [BonusItemController::class, 'update']);
+        Route::put('bonus-item/edit/{id}', [BonusItemController::class, 'update']);
         Route::delete('bonus-item/{id}', [BonusItemController::class, 'destroy']);
         Route::post('bonus-items/{id}/deliver', [BonusItemController::class, 'deliver']);
     });
 
 
+    //Item Usage Logs
     Route::prefix('item-usage-logs')->group(function () {
         Route::get('/', [ItemUsageLogController::class, 'index']);
         Route::get('/{id}', [ItemUsageLogController::class, 'show']);
-        Route::post('/', [ItemUsageLogController::class, 'store']);
+        Route::post('/create', [ItemUsageLogController::class, 'store']);
         Route::put('/{id}', [ItemUsageLogController::class, 'update']);
         Route::delete('/{id}', [ItemUsageLogController::class, 'destroy']);
         Route::get('/by-client-package/{clientPackageId}', [ItemUsageLogController::class, 'byClientPackage']);
     });
 
+    //Item Status Histories
     Route::prefix('item-status-histories')->group(function () {
         Route::post('/', [ItemStatusHistoryController::class, 'store']);
         Route::get('/{id}', [ItemStatusHistoryController::class, 'show']);
-        Route::put('/{id}', [ItemStatusHistoryController::class, 'update']);
+        Route::put('edit/{id}', [ItemStatusHistoryController::class, 'update']);
         Route::delete('/{id}', [ItemStatusHistoryController::class, 'destroy']);
         Route::get('/item/{itemId}', [ItemStatusHistoryController::class, 'indexByItem']);
     });
